Add a separate PHP 5 only file; combine in preloader
