+++ bridge/dd_init.php
@@ -3,10 +3,7 @@
 namespace DDTrace\Bridge;
 
 use DDTrace\Bootstrap;
-use DDTrace\Format;
-use DDTrace\GlobalTracer;
 use DDTrace\Integrations\IntegrationsLoader;
-use DDTrace\SpanContext;
 
 if (\PHP_VERSION_ID < 70000) {
     \date_default_timezone_set(@\date_default_timezone_get());
@@ -54,38 +51,6 @@ function dd_env_as_boolean($name, $default)
     }
 }
 
-// This gets called before curl_exec() calls from the C extension on PHP 5
-function curl_inject_distributed_headers($ch, array $headers)
-{
-    if (
-        !\class_exists('DDTrace\\SpanContext', false)
-        || !\class_exists('DDTrace\\GlobalTracer', false)
-        || !\class_exists('DDTrace\\Format', false)
-    ) {
-        return;
-    }
-    $span = GlobalTracer::get()->getActiveSpan();
-    if (null === $span) {
-        return;
-    }
-    $context = $span->getContext();
-    if (!\property_exists($context, 'origin')) {
-        return;
-    }
-
-    /*
-     * We can't use the existing context because only userland spans are represented.
-     * As a result, we create a new context with dd_trace_peek_span_id() to get the
-     * active span ID.
-     */
-    $newContext = new SpanContext($context->getTraceId(), \dd_trace_peek_span_id());
-    $newContext->origin = $context->origin;
-
-    GlobalTracer::get()->inject($newContext, Format::CURL_HTTP_HEADERS, $headers);
-
-    \curl_setopt($ch, \CURLOPT_HTTPHEADER, $headers);
-}
-
 // trigger configuration reload to memoize values of all configuration options as set by environment variables
 function_exists('dd_trace_internal_fn') && \dd_trace_internal_fn('ddtrace_reload_config');
 if (!dd_tracing_enabled()) {
@@ -95,6 +60,10 @@ if (!dd_tracing_enabled()) {
 
 // Required classes and functions
 require __DIR__ . '/autoload.php';
+if (\PHP_MAJOR_VERSION === 5) {
+    require __DIR__ . '/php5.php';
+}
+
 // Optional classes and functions
 require __DIR__ . '/dd_register_optional_deps_autoloader.php';
